# marketplace_standard_app_api/routers/system.py
from typing import Optional

from fastapi import APIRouter, HTTPException
from fastapi.responses import HTMLResponse

from ..models.system import GlobalSearchResponse

router = APIRouter(
    tags=["System"],
    responses={
        401: {"description": "Not authenticated."},
        500: {"description": "Internal server error."},
        503: {"description": "Service unavailable."},
    },
)


@router.get(
    "/globalSearch",
    operation_id="globalSearch",
    responses={
        422: {"description": "Validation error."},
        501: {"description": "Not implemented."},
    },
    response_model=GlobalSearchResponse,
)
async def global_search(
    query: str, limit: Optional[int] = 100, offset: Optional[int] = 0
) -> GlobalSearchResponse:
    """Respond to global search queries."""
    raise HTTPException(status_code=501, detail="Not implemented.")


@router.get(
    "/health",
    operation_id="heartbeat",
    response_class=HTMLResponse,
)
async def heartbeat() -> HTMLResponse:
    """Check whether the application is running and available."""
    return HTMLResponse(content="<html><body>OK</body></html>", status_code=200)